telsu: protect against too many programmes
It seems that telsu.fi now shows information for ~2.5 days when we ask
for a certain day. That triggered a failure in the change-of-day
detection code.

To protect against that we now check if we can handle another day
change. If not we abort the programme loop as we obviously have
collected anything we wanted to know about the day we asked for.